Refactor styling of version link and tag.
