+++ app/components/RuleHeader/RuleHeader.tsx
@@ -31,18 +31,6 @@ export default function RuleHeader({
   const [isEditingTitle, setIsEditingTitle] = useState(false);
   const [currTitle, setCurrTitle] = useState<string>();
   const inputRef = useRef<HTMLInputElement>(null);
-  const [currReviewBranch, setCurrReviewBranch] = useState(ruleInfo.reviewBranch);
-  const [prUrl, setPrUrl] = useState<string | null>(null);
-
-  useEffect(() => {
-    const fetchPRUrl = async () => {
-      if (currReviewBranch) {
-        const url = await getPRUrl(currReviewBranch);
-        setPrUrl(url);
-      }
-    };
-    fetchPRUrl();
-  }, [currReviewBranch]);
 
   useEffect(() => {
     const { title, filepath } = ruleInfo;
@@ -153,11 +141,6 @@ export default function RuleHeader({
               version of this rule.
             </span>
             <Flex gap="small">
-              {ruleInfo.reviewBranch && version !== RULE_VERSION.inProduction && version !== RULE_VERSION.inDev && (
-                <Button type="link" onClick={() => prUrl && window.open(prUrl, "_blank")} disabled={!prUrl}>
-                  View Pull Request
-                </Button>
-              )}
               {version !== RULE_VERSION.draft && (
                 <Button onClick={() => switchVersion(RULE_VERSION.draft)} icon={<EditOutlined />} type="dashed">
                   Draft
