+++ app/components/RuleManager/RuleManager.module.css
@@ -9,7 +9,7 @@
 
 @media (min-width: 768px) {
   .rootLayout {
-    padding-left: 7px;
-    padding-right: 7px;
+    padding-left: 12px;
+    padding-right: 12px;
   }
 }
