+++ app/components/SavePublish/SavePublish.tsx
@@ -1,10 +1,10 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import { Modal, Button, Flex, App } from "antd";
 import { SaveOutlined, UploadOutlined } from "@ant-design/icons";
 import { DecisionGraphType } from "@gorules/jdm-editor";
 import { RuleInfo } from "@/app/types/ruleInfo";
 import { updateRuleData } from "@/app/utils/api";
-import { sendRuleForReview } from "@/app/utils/githubApi";
+import { sendRuleForReview, getPRUrl } from "@/app/utils/githubApi";
 import { logError } from "@/app/utils/logger";
 import NewReviewForm from "./NewReviewForm";
 import SavePublishWarnings from "./SavePublishWarnings";
@@ -14,9 +14,10 @@ interface SavePublishProps {
   ruleInfo: RuleInfo;
   ruleContent: DecisionGraphType;
   setHasSaved: () => void;
+  version?: string | boolean;
 }
 
-export default function SavePublish({ ruleInfo, ruleContent, setHasSaved }: SavePublishProps) {
+export default function SavePublish({ ruleInfo, ruleContent, setHasSaved, version }: SavePublishProps) {
   const { _id: ruleId, filepath: filePath, reviewBranch } = ruleInfo;
 
   const { message } = App.useApp();
@@ -24,7 +25,17 @@ export default function SavePublish({ ruleInfo, ruleContent, setHasSaved }: Save
   const [currReviewBranch, setCurrReviewBranch] = useState(reviewBranch);
   const [isSaving, setIsSaving] = useState(false);
   const [isSendingToReview, setIsSendingToReview] = useState(false);
+  const [prUrl, setPrUrl] = useState<string | null>(null);
 
+  useEffect(() => {
+    const fetchPRUrl = async () => {
+      if (currReviewBranch) {
+        const url = await getPRUrl(currReviewBranch);
+        setPrUrl(url);
+      }
+    };
+    fetchPRUrl();
+  }, [currReviewBranch]);
   const save = async () => {
     setIsSaving(true);
     try {
@@ -88,18 +99,27 @@ export default function SavePublish({ ruleInfo, ruleContent, setHasSaved }: Save
         <NewReviewForm filePath={filePath} createNewReview={createNewReview} />
       </Modal>
       <Flex gap="small" justify="end" className={styles.savePublishWrapper}>
-        <Button type="primary" onClick={save} disabled={isSaving}>
-          Save <SaveOutlined />
-        </Button>
-        <Button
-          type="primary"
-          onClick={() => createOrUpdateReview()}
-          className={styles.sendForReviewBtn}
-          disabled={isSendingToReview}
-        >
-          {" "}
-          Send for Review <UploadOutlined />
-        </Button>
+        {reviewBranch && (
+          <Button type="link" onClick={() => prUrl && window.open(prUrl, "_blank")} disabled={!prUrl}>
+            View Pull Request
+          </Button>
+        )}
+        {version === "draft" && (
+          <>
+            <Button type="primary" onClick={save} disabled={isSaving}>
+              Save <SaveOutlined />
+            </Button>
+            <Button
+              type="primary"
+              onClick={() => createOrUpdateReview()}
+              className={styles.sendForReviewBtn}
+              disabled={isSendingToReview}
+            >
+              {" "}
+              Send for Review <UploadOutlined />
+            </Button>
+          </>
+        )}
       </Flex>
       <SavePublishWarnings filePath={filePath} ruleContent={ruleContent} isSaving={isSaving} />
     </>
